Move massedit/search related code to its own folder. Use this commit for diff checking.
